+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 ### Added
 - Ultrafast Grid: Obey server concurrency limitations. [Trello 2152](https://trello.com/c/yNzhBkBh)
 - Optional server timeout. [Trello 1539](https://trello.com/c/APlfd1cb)
+- Added `runner.DontCloseBatches` API in case the user doesn't want to close the batches in the end of the test. [Trello 1908](https://trello.com/c/8BGfvXKU)
 ### Updated
 - `CssParser` version updated to 1.1.0.